hardware/alsa_sound: Flush stale buffers during device switch from headset to speaker and headset combo.
- During playback over headset, launching camera and taking snapshot
results in playback audio getting mixed for sometime with shutter
sound.
- Issue is seen When camera app uses STRATEGY_ENFORCE_AUDIBLE for
shutter sound when playback is running on headset.
- With the above use case AudioFlinger mutes the playback stream
and mixes the track with shuter sound before HAL does derouting of
headset device and routing to headset+speaker combo.
- Since HAL does only derouting and routing of the device and stream
is still active, samples which are sent to firmware are played along
with shutter sound.
- Fix this issue by closing the pcm before derouting the headset
and reopening the pcm after routing to headset and speaker combo.

Change-Id: If9a44102f18e108e1de6277eb750741e91e569da
CRs-fixed: 343870